% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_test.r
\name{comp_test}
\alias{comp_test}
\title{Metric comparison test}
\usage{
comp_test(df, id_vars, ref_var, least, most)
}
\arguments{
\item{df}{Wide data frame with one row per sample. It is assumed to only 
contain one sample per sites, so if there are revisits, keep only the one 
visit. It is also assumed that all metrics are numeric, so remove any that 
are not. This data frame should include variables to identify samples and a 
single variable for level of disturbance. The data frame should include only 
the most and least disturbed classes. Some consolidation of classes may be 
desirable, depending on the situation.}

\item{id_vars}{This is a character vector with variables to identify samples
in the data frame.}

\item{ref_var}{String identifying the variable with disturbance condition 
for each site}

\item{least}{String representing value of ref_var indicating least 
disturbed condition}

\item{most}{String representing value of ref_var indicating most 
disturbed condition}
}
\value{
The output data frame includes metric, KW_stat (Kruskal-Wallis test 
statistic), KW_pval (Kruskal-Wallis test p-value), and box_score (0-3). 
The box_score represents the degree of boxplot overlap between the least 
and most disturbed groups of sites: 0 indicates overlap of median of each 
group with interquartile range (box) of other group; 
1 indicates one median does not overlap with box of other group; 2 
indicates median of each group does not overlap with box of other group, 
but boxes overlap; and 3 indicates no overlap of box of each group with 
other group. A box_score of 3 indicates the highest level of discrimination 
for a metric.
}
\description{
This function compares the most and least disturbed using a Kruskal-Wallis 
test because metrics are typically somewhat skewed. It also performs a box 
plot comparison using scoring developed by Tetra Tech for Florida SCI.
}
