% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redund_test.r
\name{redund_test}
\alias{redund_test}
\title{Metric redundancy test}
\usage{
redund_test(df, id_vars, cutoff)
}
\arguments{
\item{df}{Wide data frame with one row per sample. It is assumed to only 
contain one sample per sites, so if there are revisits, keep only the one 
visit. It is also assumed that all metrics are numeric, so remove any that 
are not.}

\item{id_vars}{This is a character vector with any non-metric variables in 
the data frame.}

\item{cutoff}{Minimum absolute correlation value at which metrics are 
flagged as redundant. The value should be a number between 0 and 1.}
}
\value{
Output dataset has `metric` with metric name, and `redund_met` is a 
string containing all metrics with an absolute correlation of at least the 
cutoff value, separated by commas.
}
\description{
This function evaluates redundancy of metrics by comparing Pearson 
correlations based on a user supplied
cutoff.
}
