% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relSOI_test.r
\name{relSOI_test}
\alias{relSOI_test}
\title{Relative scope of impairment}
\usage{
relSOI_test(df, id_vars, ref_var, least, most)
}
\arguments{
\item{df}{Wide data frame with one row per sample. It is assumed to only 
contain one sample per sites, so if there are revisits, keep only the one 
visit. It is also assumed that all metrics are numeric, so remove any that 
are not. This data frame should include variables to identify samples and a 
single variable for level of disturbance. The data frame should include only 
the most and least disturbed classes. Some consolidation of classes may be 
desirable, depending on the situation.}

\item{id_vars}{This is a character vector with variables to identify samples
in the data frame.}

\item{ref_var}{String identifying the variable with disturbance condition 
for each site}

\item{least}{String representing value of ref_var indicating least 
disturbed condition}

\item{most}{String representing value of ref_var indicating most 
disturbed condition}
}
\value{
The output data frame includes metric, least disturbed 25th 
percentile (L_q1), least disturbed 75th percentile (L_q3), least disturbed
median (L_median), most disturbed median (M_median), maximum value (max_val),
metric direction (POS/NEG, metdir), and  relative scope of impairment (soi). 
Desirable values of SOI are 1 or less. Higher values indicate excessive 
reference site variability.
}
\description{
This function calculate the relative scope of impairment of a metric, based 
on the reference designations provided by the user. Relative
scope of impairment (SOI) is calculated as the ratio of the range
of possible impairment (values beyond the poorer quality 25th
percentile) and the interquartile range (IQR) of reference sites, a measure of
variability. Relative SOI values smaller than 1 are more desirable,
with values greater than 1 indicating too much variability among
reference sites compared to the range of impairment.
}
\references{
Blocksom, KA, and BR Johnson. 2009. Development of a regional
macroinvertebrate index for large river bioassessment. Ecological 
Indicators 9:313-328.

Klemm, DJ, KA Blocksom, WT Thoeny, FA Fulk, AT Herlihy, PR Kaufmann, and
SM Cormier. 2002. Methods development and use of macroinvertebrates as 
indicators of ecological condition for streams in the Mid-Atlantic 
Highlands region. Environmental Monitoring and Assessment 78:169-212.

USEPA, 1998. Lake and Reservoir Bioassessment and Biocriteria Technical 
Guidance Document, EPA/841/B-98/007, U.S. Environmental Protection Agency, 
Office of Wetlands, Oceans, and Watersheds, Office of Science and 
Technology, Office of Water, Washington, D.C.
}
