% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_test.r
\name{sn_test}
\alias{sn_test}
\title{Signal-to-noise test}
\usage{
sn_test(df, id_vars_samp, id_vars_site, year = NULL)
}
\arguments{
\item{df}{This data frame is in wide format with one row per sample and 
assumed to contain only numeric metrics. It should also contain all visits 
to each site, and at least a subset of sites must have multiple visits. 
If there are calibration and validation subsets, this df should only contain 
calibration samples. Only identifying variables and metrics should be 
included in this data frame, and only numeric metrics can be included.}

\item{id_vars_samp}{A character vector containing variables that identify
individual samples (e.g., UID).}

\item{id_vars_site}{A string containing a variable name that 
identifies sites (e.g., SITE_ID). This cannot be the same or a subset of
variables in `id_vars_samp`. This should not be a variable that occurs across
survey cycles (e.g., UNIQUE_ID), because this would assume that variation 
across cycles is part of the noise, rather than a true change over time.}

\item{year}{A string containing the name of the year variable if sites are
revisited across years (as well as within year). The default value is NULL.}
}
\value{
Output is a data frame with the metric, signal, noise, and 
sn_ration, and comment. If the model is undefined for a given metric, the 
values for signal, noise, and sn_ratio will be NA with a comment. Otherwise,
each of these values will be provided.
}
\description{
This function calculates a version of the signal-to-noise approach used
in NARS and in EMAP. It is intended to use variation due to within-cycle 
revisits to estimate noise and variance across sites to estimate signal. 
The model used is a random effects model and assumes that sites are selected 
using a probability-based design.
}
